fix(MCRI-105) Fix sorting for Project field on indices with a projectSort field, and future proof & move the byte utility functions
